+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `continueWithAutosave` to forms validation
+
 ## [5.0.1] - 2023-11-01
 
 ## [5.0.0] - 2023-10-25