+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- public file URL for S3 assets
+
 ## [5.1.0] - 2023-11-20
 
 ### Added